pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-intake-mvp"
version = "0.1.0"
description = "Deterministic-first personal job intake system for precision remote role discovery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Codex"}
]
dependencies = [
  "beautifulsoup4>=4.12.3",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "SQLAlchemy>=2.0.36",
  "typer>=0.12.5",
  "openai>=1.52.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.7.2",
]

[project.scripts]
job-intake = "job_intake.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "C4", "SIM"]
